🎨 Improved Little Things

+++ src/app/(index)/actions.ts
@@ -1,19 +1,35 @@
 "use server";
 
+import { auth } from "@/lib/auth";
 import prisma from "@/lib/prisma";
 import { revalidatePath } from "next/cache";
 
+
 export async function handleDelete(formData: FormData) {
-    // @todo protect delete only for diary owner
-    await prisma.diary.delete({
-        where: { id: formData.get("id") as string },
-    });
-    revalidatePath("/");
+	const session = await auth();
+	if (!session) {
+		throw new Error("Unauthorized");
+	}
+	const diaryId = formData.get("id") as string;
+	await prisma.diary.delete({
+		where: { id: diaryId, accountId: session.user.sub },
+	});
+	revalidatePath("/");
 }
 
 export async function handleGoalChange(goalId: string, completed: boolean) {
-    // @todo protect diary goal change only for diary owner
-    await prisma.diaryGoal.update({
+    const session =  await auth();
+	if (!session) {
+		throw new Error("Unauthorized");
+	}
+	const goal = await prisma.diaryGoal.findUnique({
+		where: { id: goalId },
+		include: { diary: true }
+	});
+	if (goal?.diary.accountId !== session.user.sub) {
+		throw new Error("Unauthorized")
+	}
+	await prisma.diaryGoal.update({
         where: {
             id: goalId,
         },
